demo/requirements.txt
streamlit
Pillow
numpy
redisai